+++ stable/tht.h
@@ -333,6 +333,7 @@ namespace tht {
 		qRT;
 		qRE;
 		}
+		// NOTA: rearms also the blocker.
 		void Wait( void )
 		{
 		qRH
@@ -340,12 +341,10 @@ namespace tht {
 		qRB
 			Mutex.InitAndLock( Local_ );
 
-			if ( mtx::TryToLock( Main_ ) )
-				mtx::Unlock( Main_ );
-			else
-				Mutex.Unlock();
-
-			mtx::Lock( Main_ );
+			if ( !mtx::TryToLock( Main_ ) ) {
+				Mutex.reset(); // Ublocks and also avoid mutex unlocking on destruction when locked by other thread.
+        mtx::Lock(Main_);
+      }
 		qRR
 		qRT
 		qRE
@@ -364,24 +363,19 @@ namespace tht {
 		qRE
 			return Blocked;
 		}
-		void Unblock( void )
+		void Unblock(void)
 		{
 		qRH
-//			mtx::rMutex Mutex;	// Can not be used, because the destructor could be called after destruction of underlying mutes.
-			bso::sBool Locked = false;
+			mtx::rHandle Mutex;
 		qRB
-			mtx::Lock( Local_ );
-			Locked = true;
+			Mutex.InitAndLock(Local_);
 
-			if ( mtx::IsLocked( Main_ ) ) {
-				mtx::Unlock( Local_ );
-				Locked = false;
-				mtx::Unlock( Main_ );
+			if ( mtx::IsLocked(Main_) ) {
+				Mutex.reset(); // Unlocks and also avoid calling the destructor with underlying mutexes already destroyed..
+				mtx::Unlock(Main_);
 			}
 		qRR
 		qRT
-			if ( Locked )
-				mtx::Unlock( Local_ );
 		qRE
 		}
 	};