+++ tools/xdhq/wrappers/NJS/XDHqFAAS.js
@@ -108,12 +108,12 @@ const types = shared.types;
 const open = shared.open;
 const isDev = shared.isDev;
 
-const mainProtocolLabel = "22bb5d73-924f-473f-a68a-14f41d8bfa83";
-const mainProtocolVersion = "0";
+const MAIN_PROTOCOL_LABEL = "22bb5d73-924f-473f-a68a-14f41d8bfa83";
+const MAIN_PROTOCOL_VERSION = "0";
 
-const faasProtocolLabel = "4c837d30-2eb5-41af-9b3d-6c8bf01d8dbf";
-const faasProtocolVersion = "1";
-const scriptVersion ="0";
+const FAAS_PROTOCOL_LABEL = "4c837d30-2eb5-41af-9b3d-6c8bf01d8dbf";
+const FAAS_PROTOCOL_VERSION = "1";
+const SCRIPT_VERSION ="0";
 
 var token = getEnv("ATK_TOKEN");
 
@@ -361,11 +361,16 @@ const s = {
 }
 
 // Special ids.
-const ids = {
+const commandIds = {
 	FORBIDDEN: -1,
 	CREATION: -2,
 	CLOSING: -3,
 	HEAD: -4,
+}
+
+const responseIds = {
+	BROADCAST: -3,
+	HEAD: -4
 }
 
 function handleCommand(command, head) {
@@ -374,19 +379,19 @@ function handleCommand(command, head) {
 	// console.log(command);
 
 	switch (command) {
-	case ids.FORBIDDEN:
+	case commandIds.FORBIDDEN:
 		exit_("Received unexpected undefined command id!");
 		break;
-	case ids.CREATION:
+	case commandIds.CREATION:
 		push(s.CREATION);
 		push(d.SINT);
 		break;
-	case ids.CLOSING:
+	case commandIds.CLOSING:
 		push(s.CLOSING);
 		push(d.SINT);
 		break;
-	case ids.HEAD:
-		socket.write(addString(convertSInt(-4), head === undefined ? "" : head));
+	case commandIds.HEAD:
+		socket.write(addString(convertSInt(responseIds.HEAD), head === undefined ? "" : head));
 		break;
 	default:
 		if (command < 0 )
@@ -695,7 +700,7 @@ function handshakes(feeder) {
 			if ( string.length )
 				process.stdout.write(string + '\n');
 
-			socket.write(addString(addString(handleString(mainProtocolLabel),mainProtocolVersion),scriptVersion));
+			socket.write(addString(addString(handleString(MAIN_PROTOCOL_LABEL),MAIN_PROTOCOL_VERSION),SCRIPT_VERSION));
 			pop();
 			push(h.ERROR_MAIN);
 			push(d.STRING);
@@ -771,7 +776,7 @@ function launch(createCallback, actionCallbacks, head) {
 		push(d.STRING);
 		socket.on('data', (data) => onRead(data, createCallback, actionCallbacks, head));
 		
-		socket.write(addString(addString(handleString(faasProtocolLabel),faasProtocolVersion),"NJS " + version));
+		socket.write(addString(addString(handleString(FAAS_PROTOCOL_LABEL),FAAS_PROTOCOL_VERSION),"NJS " + version));
 	});	
 }
 
@@ -827,7 +832,7 @@ function broadcastAction(action, id) {
 	if ( ( action === undefined ) || ( action === "" ) )
 		exit_("There must be an non-empty action parameter for tha broadcastAction function!");
 
-	socket.write(addString(addString(convertSInt(-3), action), id === undefined ? "" : id ));
+	socket.write(addString(addString(convertSInt(responseIds.BROADCAST), action), id === undefined ? "" : id ));
 }
 
 module.exports.launch = launch;